Replaced "application/json" with ContentType.APPLICATION_JSON (#64)
* Added model in client, added ChatService and EventBusService

* Changed test connection url

* Changed Input with ng model

* Extracted register and login forms

* Continued implementation

* Resolved errors

* Resolved error in AutenticationDataStoreDatabase

* Resolved all errors

* Added InsertParams to tracked files

* Added Eclipse gitignore

* Added Vert.x output folder to gitignore

* Started to test interaction between client and server

* Recompiled custom.css and custom.min.css

* Made server interceptor more generale

* Fixed register

* Login works

* Fixed add-room

* Ranamed variable

* Fixed closing of add-room

* Replaced ' with "

* Removed char

* Added authentication checks in login, register and homepage

* Logout works

* Removed unused import

* Added custom.min.css

* Resolved error in user login

* Removed template and added material layout

* Saving changes before commit

* Resolved merge conflicts

* Removed print of port in run.sh

* Resolved token error and room web client retrieval

* Implemented Vert.x's Event Bus address abstraction

* Fixed compile errors

* Verified that npm install does not throw an error

* Changed method observableRule overloading addRule

* Register works

* Removed Eclipse project files

* Made files executable

* Fixed new lines

* Resolved compilation errors

* Made gradlew scripts executable

* Fixed compilation error

* Upgraded Vert.x version to 3.5.4

* Added consumer, publisher and subscriber abstraction

* Adding changes

* Fixed error, room creation should not be broadcasted

* Replaced "application/json" with dedicated content type

* Fixed errors in Angular build script

* Optimized imports

+++ dcs-client/build.sh
@@ -1,5 +1,7 @@
 #!/bin/bash
 
+set -e
+
 SRC=./dist
 DEST=../dcs-webapp-service/src/main/resources/webroot
 
@@ -12,8 +14,10 @@ echo "Building Angular application..."
 ng build --prod --aot
 
 if [ -d "$DEST" ]; then
-    echo "Removing contents from folder '$DEST'..."
-    rm -r $DEST/*
+    if [ `ls -1A $DEST | wc -l` -ne 0 ]; then
+        echo "Removing contents from folder '$DEST'..."
+        rm -r $DEST/*
+    fi
 else
     echo "Creating folder '$DEST'..."
     mkdir -p $DEST
